Started on my Peep Tracker
You remade everything from scratch. You decided to add another category for the "peeps". Unfortunately, that meant that you had to the delete the table so that you could add the new column with all the correct attributes and WITHOUT sql screaming at you.

You have completed the HTML elements. They should all have the correct attributes. They just need to be given functions via the script.js.

The script.js has yet to be written.

The CSS has yet to be written.