Welp, you really tried
So, you added a database for the user and a database for their peeps. You managed to figure out how to get all that connected to a server (with a lot of help), you started to understand about APIs a little more. But not nearly enough to actually have this project completed by tomorrow. You also finished your sequence diagram. Sort of.